requirements.txt
asgiref==3.3.1
Django==3.1.4
django-appconf==1.0.4
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Pillow==8.0.1
psycopg2-binary==2.8.6
PyJWT==2.0.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
